Fix annoying warnings coming from osips_parser

+++ lib/osips_parser/ut.h
@@ -1011,7 +1011,7 @@ static inline int str_strcmp(const str *stra, const str *strb)
  */
 static inline int str_match_nt(const str *a, const char *b)
 {
-	return a->len == strlen(b) && !memcmp(a->s, b, a->len);
+	return a->len == (int)strlen(b) && !memcmp(a->s, b, a->len);
 }
 
 /*
@@ -1019,7 +1019,7 @@ static inline int str_match_nt(const str *a, const char *b)
  */
 static inline int str_casematch_nt(const str *a, const char *b)
 {
-	return a->len == strlen(b) && !strncasecmp(a->s, b, a->len);
+	return a->len == (int)strlen(b) && !strncasecmp(a->s, b, a->len);
 }
